fix: class 컴포넌트 삭제
