+++ just-swim-app/src/app/(afterLogin)/instructor/class/classView.module.scss
